makes share functionality not broken
